repoproj/creditapproval/requirements.txt
Django==3.2.5
djangorestframework==3.12.4
openpyxl==3.0.7
